server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.120.4",
    "httpx>=0.28.1",
    "beautifulsoup4>=4.12.3",
    "pyyaml>=6.0.1",
    "uvicorn>=0.38.0",
    "pyaml>=25.7.0",
]
